requirements.txt
python-dotenv
langchain
openai
requests